requirements.txt
# Data preparation dependencies
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Note: YOLOv7 dependencies should be installed separately
# See: https://github.com/WongKinYiu/yolov7#installation